Order feature and Error handling

+++ src/main/java/com/bookstore/resource/AuthorResource.java
@@ -4,21 +4,15 @@ import java.net.URI;
 import java.util.List;
 
 import com.bookstore.exception.AuthorNotFoundException;
+import com.bookstore.exception.InvalidInputException;
 import com.bookstore.model.Author;
 import com.bookstore.model.Book;
 import com.bookstore.repo.AuthorRepository;
 import com.bookstore.repo.BookRepository;
 
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.DELETE;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.PUT;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.PathParam;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.Response;
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 
 @Path("/authors")
 @Produces(MediaType.APPLICATION_JSON)
@@ -29,6 +23,7 @@ public class AuthorResource {
 
     @POST
     public Response createAuthor(Author author){
+        validateAuthor(author);
         Author createdAuthor = authorRepository.save(author);
         return Response.created(URI.create("/authors/"+ createdAuthor.getId()))
         .entity(createdAuthor)
@@ -44,6 +39,9 @@ public class AuthorResource {
     @Path("/{id}")
     public Author findById(@PathParam("id") Long id){
         Author author = authorRepository.findById(id);
+        if(id == null) {
+            throw new InvalidInputException("Author id cannot be null");
+        }
         if (author == null) {
             throw new AuthorNotFoundException("Author with id "+ id + " not found.");
         }
@@ -53,6 +51,9 @@ public class AuthorResource {
     @PUT
     @Path("/{id}")
     public Author updateAuthor (@PathParam("id") Long id,Author author){
+        if (id == null) {
+            throw new InvalidInputException("Author id cannot be null"); 
+        }
         if (authorRepository.findById(id) == null) {
             throw new AuthorNotFoundException("Author with id "+ id + " not found.");
         }
@@ -63,22 +64,35 @@ public class AuthorResource {
     @DELETE
     @Path("/{id}")
     public Response deleteAuthor(@PathParam("id") Long id){
+        if (id == null) {
+            throw new InvalidInputException("Author id cannot be null"); 
+        }
         if (authorRepository.findById(id) == null) {
             throw new AuthorNotFoundException("Author with id "+ id + " not found.");
         }
         authorRepository.delete(id);
-        return Response.ok()
-        .entity("Author with id " + id + " deleted successfully.")
-        .build();
+        return Response.noContent().build();
     }
 
     @GET
     @Path("/{id}/books")
     public List<Book> getBooksByAuthorId(@PathParam("id") Long id){
+        if (id == null) {
+            throw new InvalidInputException("Author id cannot be null"); 
+        }
         if (authorRepository.findById(id) == null) {
             throw new AuthorNotFoundException("Author with id "+ id + " not found.");
         }
         return bookRepository.findByAuthorId(id);
     }
-    
+
+    // Helper method for validation Author
+    private void validateAuthor(Author author) {
+        if (author == null) {
+            throw new InvalidInputException("Author cannot be null");
+        }
+        if (author.getName() == null || author.getName().isEmpty()) {
+            throw new InvalidInputException("Author name cannot be null or empty");
+        }
+    }
 }
